chore(mangrove.js): move test utils to util/test for clarity & easier later browser omission
