+++ packages/mangrove.js/test/integration/semibook.integration.test.ts
@@ -2,7 +2,7 @@
 import { describe, beforeEach, afterEach, it } from "mocha";
 import { expect } from "chai";
 
-import * as mgvTestUtil from "../../src/util/mgvIntegrationTestUtil";
+import * as mgvTestUtil from "../../src/util/test/mgvIntegrationTestUtil";
 const waitForTransaction = mgvTestUtil.waitForTransaction;
 import { newOffer, toWei } from "../util/helpers";
